Update patch set 4
Patch Set 4: Code-Review-1

(8 comments)

Another comment: Why are all the atl-lab1 deployment files contained in this patchset? Those files are unrelated to the change of allowing linting policies to be specified. If we've decided to add those deployment files to Pegleg, that should be done in a separate patchset.

Patch-set: 4
Label: Code-Review=-1