Add top-level lint/fmt tox jobs
This is to add a top-level lint/fmt job to Pegleg so that tox -e lint
is executed in CICD. lint is included under envlist to accomplish this.
fmt is excluded from envlist as it doesn't need to be executed in CICD
but is included nonetheless for convenience.

Also fixes some typos.

Change-Id: I76165704b32653c98e506ddde71c8240d5f28492